Re-factor provider structure
Change providers to be node and database providers.
This allows for the cration of databases in the future
