+++ src/pages/Create.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Create: React.FC = () => {
+  return <div>Create</div>
+}
+
+export default Create;
